Create index_script
This script will import and download the necessary libraries needed to index the books.  You may only have one large book or multiple books which is typical.  This will not work if you have not "removed" the password from your PDF books.  There is a script available to help you with that effort.